src/buffer.rs: Consolidate from Corroded `buffer*.rs`
Consolidates all Corrode-translated source code for buffer-related
functionality into a common `buffer` module.

Moves all `buffer_*()` into an `impl buffer` block, as these functions
should eventually all accept `self`.

Adds a `type Op` aliased to the C function pointer type for buffer
operations, as otherwise its signature is a bit unwieldy.

Code is otherwise unchanged from the original Corrode output.
